// src/api/__generated__/wallet/model/blockRelationResponseRelation.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 当前登录用户与目标用户的拉黑关系
 */
export type BlockRelationResponseRelation =
  (typeof BlockRelationResponseRelation)[keyof typeof BlockRelationResponseRelation];

export const BlockRelationResponseRelation = {
  NONE: 'NONE',
  BLOCKED_BY_ME: 'BLOCKED_BY_ME',
  BLOCKED_BY_OTHER: 'BLOCKED_BY_OTHER',
  MUTUAL_BLOCK: 'MUTUAL_BLOCK',
} as const;